+++ app/controllers/federails/client_controller.rb
@@ -1,5 +1,5 @@
 module Federails
-  class ClientController < ActionController::Base
+  class ClientController < Federails.configuration.base_client_controller.constantize
     include Pundit::Authorization
 
     layout Federails.configuration.app_layout if Federails.configuration.app_layout