+++ CHANGELOG.md
@@ -28,6 +28,11 @@ Please, keep them in this order when updating.
 
 ## [Unreleased]
 
+### Added
+
+- Base controller for client controllers can be specified to something different from `ActionController::Base` with the 
+  `base_client_controller` option
+
 ## [0.2.0] 2024-10-13 - Sign messages, handle signed messages and remote following
 
 ### Added